+++ ocrd/bashlib/src/lib.bash
@@ -1,3 +1,4 @@
+# HERE-INCLUDE ./src/bash_version_check.bash
 # HERE-INCLUDE ./src/raise.bash
 # HERE-INCLUDE ./src/dumpjson.bash
 # HERE-INCLUDE ./src/usage.bash