+++ ocrd/ocrd/lib.bash
@@ -1,10 +1,11 @@
+# BEGIN-INCLUDE ./src/bash_version_check.bash 
+((BASH_VERSINFO<4 || BASH_VERSINFO==4 && BASH_VERSINFO[1]<4)) && \
+    echo >&2 "bash $BASH_VERSION is too old. Please install bash 4.4 or newer." && \
+    exit 1
+
+# END-INCLUDE 
 # BEGIN-INCLUDE ./src/raise.bash 
-## ### `ocrd__raise`
-## 
-## Raise an error and exit.
-ocrd__raise () {
-    echo >&2 "ERROR: $1"; exit 127
-}
+
 
 # END-INCLUDE 
 # BEGIN-INCLUDE ./src/dumpjson.bash