+++ tests/processor/test_processor.py
@@ -28,6 +28,19 @@ class DummyProcessor(Processor):
         kwargs['version'] = '0.0.1'
         super(DummyProcessor, self).__init__(*args, **kwargs)
 
+class DummyProcessorWithRequiredParameters(Processor):
+    def process(self): pass
+    def __init__(self, *args, **kwargs):
+        kwargs['version'] = '0.0.1'
+        kwargs['ocrd_tool'] = {
+            'executable': 'ocrd-test',
+            'steps': ['recognition/post-correction'],
+            'parameters': {
+                'i-am-required': {'required': True}
+            }
+        }
+        super(DummyProcessorWithRequiredParameters, self).__init__(*args, **kwargs)
+
 class IncompleteProcessor(Processor):
     pass
 
@@ -76,18 +89,6 @@ class TestProcessor(TestCase):
         DummyProcessor(self.workspace, dump_json=True)
 
     def test_params_missing_required(self):
-        class DummyProcessorWithRequiredParameters(Processor):
-            def process(self): pass
-            def __init__(self, *args, **kwargs):
-                kwargs['version'] = '0.0.1'
-                kwargs['ocrd_tool'] = {
-                    'executable': 'ocrd-test',
-                    'steps': ['recognition/post-correction'],
-                    'parameters': {
-                        'i-am-required': {'required': True}
-                    }
-                }
-                super(DummyProcessorWithRequiredParameters, self).__init__(*args, **kwargs)
         with self.assertRaisesRegex(Exception, 'is a required property'):
             DummyProcessorWithRequiredParameters(workspace=self.workspace)
 
@@ -103,6 +104,7 @@ class TestProcessor(TestCase):
 
     def test_run_cli(self):
         with TemporaryDirectory() as tempdir:
+            run_processor(DummyProcessor, ocrd_tool=DUMMY_TOOL, workspace=self.workspace)
             run_cli(
                 'echo',
                 mets_url=assets.url_of('SBB0000F29300010000/data/mets.xml'),